deleted App.test.js file
