+++ .babelrc
@@ -1,3 +1,13 @@
 {
-  "presets": ['@babel/preset-env', '@babel/preset-react']
+  "presets": [
+    [
+      "@babel/preset-env",
+      {
+        "targets": {
+          "node": "10"
+        }
+      }
+    ],
+    "@babel/preset-react"
+  ]
 }
