+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "jest",
     "react-dev": "webpack -d --watch",
-    "server-dev": "nodemon server/index.js",
+    "server-dev": "nodemon ./server/server.js",
     "seed": "node database/seed.js"
   },
   "keywords": [],
@@ -28,6 +28,7 @@
     "jest": "^24.8.0",
     "npm": "^6.9.0",
     "react-test-renderer": "^15.6.2",
+    "supertest": "^4.0.2",
     "webpack": "^4.31.0",
     "webpack-cli": "^3.3.2"
   },
