+++ server/index.js
@@ -1,15 +1,13 @@
 const express = require('express');
-const mongoose = require('mongoose');
-const House = require('../database/db.js');
+const db = require('../database/db.js');
 
 const app = express();
-const port = 3002;
 
 app.use(express.static(`${__dirname}/../client/dist`));
 
 app.get('/house/:houseId', (req, res) => {
   console.log(req.params)
-  House.findOne({ house_id: req.params.houseId }, (err, house) => {
+  db.House.findOne({ house_id: req.params.houseId }, (err, house) => {
     if (err) {
       res.status(404);
       res.send();
@@ -20,4 +18,5 @@ app.get('/house/:houseId', (req, res) => {
   });
 });
 
-app.listen(port, () => console.log(`App listening on port ${port}!`));
+
+module.exports = app;
