+++ server/server.js
@@ -1,0 +1,5 @@
+const app = require('./index.js');
+
+const port = 3002;
+
+app.listen(port, () => console.log(`App listening on port ${port}!`));
