+++ tests/enzyme.config.js
